requirements.txt
PyQt5==5.15.9
PyQt5-sip==12.13.0
Pillow>=10.0.0